Create ENGR 132 prototype code (group 2)
First thing the device will turn on and start scanning the tag. The next step is detecting if the tag is the correct or not. If it is the correct tag, it should give access and if not, it must go back to the start point. If the tag access is accessible, then the lock should be unlocked and then wait for 10 seconds then lock again and if not, the lock must remain locked.